Improved the test that checks the default post monetization settings are retrieved correctly from the database. Fixed spelling errors.
